Phase 8: display layer (titles, action bar, boss bar, particles, map, home)
User-facing:
- Border title on chunk-cross when ownership changes — 'Entering <Realm>'
  / 'Wilderness', color-coded by relation (own/ally/enemy/neutral/peaceful).
  Subtitle shows mayor + chunk count, [Peaceful], or [Weakened]. Optional
  border-cross sound (off by default).
- Persistent action bar (default) shows current chunk owner, relation tag
  ((Ally)/(Enemy)), and [Peaceful]/[Weakened] tags.
- Opt-in boss bar — color-coded by relation; fill = 1 - claim_cost/capacity
  (near-empty → about to be weakened).
- /realm showclaim [line|wall|corner|off]: clientside particle outline of
  nearby ownership boundaries (radius 5). Auto-pauses below TPS threshold.
- /realm map: 11×11 ASCII grid, colored by relation; legend included.
- /realm display title|bar|sound + /realm togglebar for per-player prefs
  (persisted in display_prefs table on first change).
- /realm sethome (mayor, must stand in own claim) + /realm home / spawn
  (warmup + cooldown, cancel-on-move and cancel-on-damage).

Internals:
- Palette resolves relation → legacy color code + RGB Color.
- BorderTitleListener: chunk-cross filter via PlayerMoveEvent comparing
  from/to chunks, plus PlayerChangedWorldEvent. lastOwner map seeds on
  join, clears on quit.
- TerritoryDisplayTask: 1Hz repeating BukkitRunnable; ACTION mode sends
  action bar packets, BOSS mode maintains a per-player BossBar handle in
  liveBars and updates it in place.
- ShowClaimManager: 1Hz BukkitRunnable; iterates radius×radius chunks per
  visualizer and only draws south + east edges to avoid double-render.
  Particles via Player.spawnParticle (clientside). TPS guard pauses
  rendering globally when TPS < threshold.
- HomeManager: warmup → re-fetch fresh home in the lambda before
  teleporting (prevents stale-snapshot exploit if home is cleared during
  warmup), set cooldown only on success. cancel-on-move uses block-level
  comparison so head-turn / sneak don't cancel.
- DisplayQuitListener: drops boss-bar handles + showclaim slots on
  PlayerQuitEvent.

Review-fix highlights (caught before commit):
- THREE Long==long auto-box identity comparisons fixed (Palette.relationFor,
  HomeManager.setHome, OverclaimManager.start). Java's Long cache only
  covers -128..127, so any realm with id≥128 silently saw 'OWN'/'self-chunk'
  comparisons fail. This affected EVERY server with more than ~127 realms.
- Boss-bar leak on quit (TerritoryDisplayTask had an onPlayerQuit method
  that was never wired). New DisplayQuitListener calls it.
- ShowClaim slot leak on quit — same fix path.
- Boss bar always re-shown each tick so reconnects with stale handles
  don't render to a dead session.
- Home warmup re-fetches the realm + home location inside the runTaskLater
  lambda so a mid-warmup home-clear or disband cancels the teleport.

+++ realms/src/main/java/com/realms/manager/OverclaimManager.java
@@ -81,7 +81,9 @@ public final class OverclaimManager {
 
         ClaimKey here = ClaimKey.of(player.getLocation());
         Long ownerId = store.claimOwner(here);
-        if (ownerId == null || ownerId == myRealm.id()) {
+        // Long == long auto-boxes — identity comparison only works for cached
+        // ids (-128..127). Use longValue so ids above 127 also resolve.
+        if (ownerId == null || ownerId.longValue() == myRealm.id()) {
             return Result.fail("errors.chunk-not-overclaimable");
         }
         Realm victim = store.getRealm(ownerId);
